requirements.txt
appdirs==1.4.3
attrs==19.3.0
biopython==1.76
black==19.10b0
click==7.1.1
cycler==0.10.0
decorator==4.4.2
kiwisolver==1.2.0
lxml==4.5.0
matplotlib==3.2.2
more-itertools==8.4.0
networkx==2.4
numpy==1.18.2
packaging==20.4
pathspec==0.7.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pyteomics==4.2
pytest==5.4.3
pytest-datadir==1.3.1
python-dateutil==2.8.1
regex==2020.2.20
six==1.15.0
toml==0.10.0
typed-ast==1.4.1
wcwidth==0.2.5
